Move render test to test_bridges_env and instantiate class state attribute with None
